Add files via upload
Live MATLAB interactive report of data science questions. The report included 6 questions from 3 different categories, with each question having 5-10 variants generated using MATLAB.
This course project is part of the Mathematical Foundations of Data Science course. The questions are based on the mathematical concepts that are used in data science.